Foreign Key Constraints ✨ (#5615)
* Bump knex-schema-inspector

* Fix cli role name attr

* Update relation type

* Restructure relations

* Restructure relations table

* Update api type for relation record

* Fetch relations in new structure

* Update schema-inspector

* Use new relations schema structure in api

* Update relations GETters

* Add default value to one deselect

* Add create relationship on existing field

* Add updating existing relationship

* Add delete relations

* Add relations query resolver

* Add graphql mutations for relations

* Fix reading from wrong name

* Fix wrong method name

* No idea why this flip flops every install

* Update relation type

* Accept null in use-collection composable

* Use new relations structure in translations

* Use new relations structure in new-collection

* Start updating field detail store

* Renames for new relations structure

* Silently ignore passed collection/field in relation update

* Fix setting pk field in m2o relational setup

* Small tweaks in o2m setup

* Fix m2m setup

* Tweak m2o setup

* Fix m2a setup

* Allow null for related collection (m2a)

* Fix languages code name

* Fix migration default value

* Fix relational cleanup in collections/fields

* Fix transaction problem in field delete

* Fix inserting relational o2m items

* Don't execute updateByQuery on empty item set

Fixes #5710, fixes #5070

* Show referential action input on m2o

* Finish language for m2o

* Show triggers config on o2m

* Delete items on one_deselect_item delete

* Fix naming, show relational trigger config on m2m

* Tweak language, add setup to m2a

* Fix linter warnings

* Add trigger setup for translations

* fix Edit non-schema relationship issue

* Sync existing on_delete triggers in o2m setup

* Add migration to setup foreign key constraints

* Update illegal FK values before setting constraint

* Fix MySQL unsigned vs not-unsigned in FK creation

* Use pretty names for labels in relational triggers

* Prefix auto-junction when system table

Fixes #5493

* Add system foreign key triggers

Fixes #5749

* Update docs

+++ api/src/services/collections.ts
@@ -339,54 +339,69 @@ export class CollectionsService {
 			throw new ForbiddenException();
 		}
 
-		const collectionItemsService = new ItemsService('directus_collections', {
-			knex: this.knex,
-			accountability: this.accountability,
-			schema: this.schema,
-		});
-
-		const fieldsService = new FieldsService({
-			knex: this.knex,
-			accountability: this.accountability,
-			schema: this.schema,
-		});
-
 		const tablesInDatabase = Object.keys(this.schema.collections);
 
 		if (tablesInDatabase.includes(collectionKey) === false) {
 			throw new ForbiddenException();
 		}
 
-		await collectionItemsService.deleteOne(collectionKey);
+		await this.knex.transaction(async (trx) => {
+			const collectionItemsService = new ItemsService('directus_collections', {
+				knex: trx,
+				accountability: this.accountability,
+				schema: this.schema,
+			});
 
-		await this.knex('directus_fields').delete().where('collection', '=', collectionKey);
-		await this.knex('directus_presets').delete().where('collection', '=', collectionKey);
-		await this.knex('directus_revisions').delete().where('collection', '=', collectionKey);
-		await this.knex('directus_activity').delete().where('collection', '=', collectionKey);
-		await this.knex('directus_permissions').delete().where('collection', '=', collectionKey);
+			const fieldsService = new FieldsService({
+				knex: trx,
+				accountability: this.accountability,
+				schema: this.schema,
+			});
 
-		const relations = this.schema.relations.filter((relation) => {
-			return relation.many_collection === collectionKey || relation.one_collection === collectionKey;
-		});
+			await trx('directus_fields').delete().where('collection', '=', collectionKey);
+			await trx('directus_presets').delete().where('collection', '=', collectionKey);
 
-		for (const relation of relations) {
-			const isM2O = relation.many_collection === collectionKey;
+			const revisionsToDelete = await trx.select('id').from('directus_revisions').where({ collection: collectionKey });
 
-			if (isM2O) {
-				await this.knex('directus_relations')
-					.delete()
-					.where({ many_collection: collectionKey, many_field: relation.many_field });
-
-				await fieldsService.deleteField(relation.one_collection!, relation.one_field!);
-			} else if (relation.one_collection) {
-				await this.knex('directus_relations')
-					.update({ one_field: null })
-					.where({ one_collection: collectionKey, one_field: relation.one_field });
-				await fieldsService.deleteField(relation.many_collection, relation.many_field);
+			if (revisionsToDelete.length > 0) {
+				const keys = revisionsToDelete.map((record) => record.id);
+				await trx('directus_revisions').update({ parent: null }).whereIn('parent', keys);
 			}
-		}
 
-		await this.knex.schema.dropTable(collectionKey);
+			await trx('directus_revisions').delete().where('collection', '=', collectionKey);
+
+			await trx('directus_activity').delete().where('collection', '=', collectionKey);
+			await trx('directus_permissions').delete().where('collection', '=', collectionKey);
+			await trx('directus_relations').delete().where({ many_collection: collectionKey });
+
+			const relations = this.schema.relations.filter((relation) => {
+				return relation.collection === collectionKey || relation.related_collection === collectionKey;
+			});
+
+			for (const relation of relations) {
+				// Delete related o2m fields that point to current collection
+				if (relation.related_collection && relation.meta?.one_field) {
+					await fieldsService.deleteField(relation.related_collection, relation.meta.one_field);
+				}
+
+				// Delete related m2o fields that point to current collection
+				if (relation.related_collection === collectionKey) {
+					await fieldsService.deleteField(relation.collection, relation.field);
+				}
+
+				const isM2O = relation.collection === collectionKey;
+
+				// Delete any fields that have a relationship to/from the current collection
+				if (isM2O && relation.related_collection && relation.meta?.one_field) {
+					await fieldsService.deleteField(relation.related_collection!, relation.meta.one_field);
+				} else {
+					await fieldsService.deleteField(relation.collection, relation.field);
+				}
+			}
+
+			await collectionItemsService.deleteOne(collectionKey);
+			await trx.schema.dropTable(collectionKey);
+		});
 
 		if (cache && env.CACHE_AUTO_PURGE && opts?.autoPurgeCache !== false) {
 			await cache.clear();
