+++ api/src/services/items.ts
@@ -589,6 +589,9 @@ export class ItemsService<Item extends AnyItem = AnyItem> implements AbstractSer
 
 		const itemsToDelete = await itemsService.readByQuery(readQuery);
 		const keys: PrimaryKey[] = itemsToDelete.map((item: AnyItem) => item[primaryKeyField]);
+
+		if (keys.length === 0) return [];
+
 		return await this.deleteMany(keys);
 	}
